06-user-service/app/core/service_discovery.py
"""
================================================================================
FILE IDENTITY
================================================================================
Project      : Gravity MicroServices Platform - User Service
File         : app/core/service_discovery.py
Description  : Consul service discovery integration
Language     : English (UK)
Framework    : FastAPI / Python 3.11+ / Consul

================================================================================
AUTHORSHIP & CONTRIBUTION
================================================================================
Primary Author    : GitHub Copilot
Contributors      : Elena Volkov (Backend & Integration Lead)
Team Standard     : Elite Engineers (IQ 180+, 15+ years experience)

================================================================================
TIMELINE & EFFORT
================================================================================
Created Date      : 2025-11-08
Last Modified     : 2025-11-08
Development Time  : 1 hour
Total Cost        : 1 × $150 = $150.00 USD

================================================================================
VERSION HISTORY
================================================================================
v1.0.0 - 2025-11-08 - GitHub Copilot - Initial service discovery integration

================================================================================
DEPENDENCIES
================================================================================
Internal  : app.config
External  : python-consul, httpx
Service   : Consul 1.16+

================================================================================
LICENSE & COPYRIGHT
================================================================================
Copyright (c) 2025 Gravity MicroServices Platform
License: MIT License
Repository: https://github.com/GravityWavesMl/user-service

================================================================================
"""

import logging
from typing import Optional
import consul
import httpx
from app.config import settings

logger = logging.getLogger(__name__)


class ServiceDiscovery:
    """
    Consul service discovery client.
    
    Handles:
    - Service registration
    - Health check registration
    - Service deregistration
    - Service lookup
    """
    
    def __init__(self):
        """Initialize Consul client."""
        # Parse Consul URL
        consul_url = settings.SERVICE_DISCOVERY_URL.replace("http://", "").replace("https://", "")
        if ":" in consul_url:
            host, port = consul_url.split(":")
            self.consul = consul.Consul(host=host, port=int(port))
        else:
            self.consul = consul.Consul(host=consul_url, port=8500)
        
        self.service_id = settings.SERVICE_ID
        self.service_name = settings.SERVICE_NAME
        self.service_port = settings.PORT
        self.service_host = settings.HOST if settings.HOST != "0.0.0.0" else "localhost"
    
    def register(self) -> bool:
        """
        Register service with Consul.
        
        Returns:
            bool: True if registration successful, False otherwise
        """
        try:
            # Health check configuration
            health_check = consul.Check.http(
                url=f"http://{self.service_host}:{self.service_port}/health",
                interval="10s",
                timeout="5s",
                deregister="30s"  # Auto-deregister after 30s if unhealthy
            )
            
            # Register service
            self.consul.agent.service.register(
                name=self.service_name,
                service_id=self.service_id,
                address=self.service_host,
                port=self.service_port,
                tags=[
                    "user-service",
                    "v1",
                    f"version-{settings.API_VERSION}",
                    "http",
                    "rest"
                ],
                check=health_check,
                meta={
                    "version": settings.API_VERSION,
                    "description": settings.API_DESCRIPTION,
                }
            )
            
            logger.info(
                f"✅ Service registered with Consul: "
                f"{self.service_name} ({self.service_id}) "
                f"at {self.service_host}:{self.service_port}"
            )
            return True
            
        except Exception as e:
            logger.error(f"❌ Failed to register service with Consul: {e}")
            return False
    
    def deregister(self) -> bool:
        """
        Deregister service from Consul.
        
        Returns:
            bool: True if deregistration successful, False otherwise
        """
        try:
            self.consul.agent.service.deregister(self.service_id)
            logger.info(f"✅ Service deregistered from Consul: {self.service_id}")
            return True
            
        except Exception as e:
            logger.error(f"❌ Failed to deregister service from Consul: {e}")
            return False
    
    def get_service(self, service_name: str) -> Optional[dict]:
        """
        Get service information from Consul.
        
        Args:
            service_name: Name of service to lookup
        
        Returns:
            dict: Service information with host and port, or None if not found
        """
        try:
            # Get healthy service instances
            _, services = self.consul.health.service(service_name, passing=True)
            
            if not services:
                logger.warning(f"⚠️ No healthy instances found for service: {service_name}")
                return None
            
            # Return first healthy instance
            service = services[0]
            service_info = {
                "name": service_name,
                "host": service["Service"]["Address"],
                "port": service["Service"]["Port"],
                "service_id": service["Service"]["ID"],
                "tags": service["Service"]["Tags"],
                "meta": service["Service"]["Meta"]
            }
            
            logger.info(f"✅ Found service: {service_name} at {service_info['host']}:{service_info['port']}")
            return service_info
            
        except Exception as e:
            logger.error(f"❌ Failed to get service from Consul: {e}")
            return None
    
    async def get_service_url(self, service_name: str) -> Optional[str]:
        """
        Get full service URL from Consul.
        
        Args:
            service_name: Name of service to lookup
        
        Returns:
            str: Full service URL (e.g., http://localhost:8081), or None if not found
        """
        service_info = self.get_service(service_name)
        if service_info:
            return f"http://{service_info['host']}:{service_info['port']}"
        return None
    
    def health_check(self) -> dict:
        """
        Perform health check.
        
        Returns:
            dict: Health check result
        """
        try:
            # Check Consul connectivity
            leader = self.consul.status.leader()
            
            return {
                "status": "healthy",
                "consul_connected": True,
                "consul_leader": leader,
                "service_id": self.service_id,
                "service_name": self.service_name
            }
            
        except Exception as e:
            logger.error(f"❌ Health check failed: {e}")
            return {
                "status": "unhealthy",
                "consul_connected": False,
                "error": str(e)
            }


# Global service discovery instance
service_discovery: Optional[ServiceDiscovery] = None


def get_service_discovery() -> ServiceDiscovery:
    """
    Get global service discovery instance.
    
    Returns:
        ServiceDiscovery: Service discovery instance
    """
    global service_discovery
    if service_discovery is None:
        service_discovery = ServiceDiscovery()
    return service_discovery


async def register_service() -> bool:
    """
    Register service with Consul on startup.
    
    Returns:
        bool: True if registration successful
    """
    try:
        sd = get_service_discovery()
        return sd.register()
    except Exception as e:
        logger.error(f"❌ Failed to register service: {e}")
        return False


async def deregister_service() -> bool:
    """
    Deregister service from Consul on shutdown.
    
    Returns:
        bool: True if deregistration successful
    """
    try:
        sd = get_service_discovery()
        return sd.deregister()
    except Exception as e:
        logger.error(f"❌ Failed to deregister service: {e}")
        return False


async def get_auth_service_url() -> str:
    """
    Get Auth Service URL from Consul or fallback to config.
    
    Returns:
        str: Auth Service URL
    """
    try:
        sd = get_service_discovery()
        url = await sd.get_service_url("auth-service")
        if url:
            return url
        
        # Fallback to config
        logger.warning("⚠️ Using Auth Service URL from config (Consul not available)")
        return settings.AUTH_SERVICE_URL
        
    except Exception as e:
        logger.error(f"❌ Failed to get Auth Service URL: {e}")
        return settings.AUTH_SERVICE_URL
