06-user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "1.0.0"
description = "User profile and preference management service"
authors = ["Elena Volkov <elena.volkov@gravity.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "~3.12.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = {extras = ["hiredis"], version = "^5.0.1"}
httpx = "^0.25.2"
pillow = "^10.1.0"
python-magic = "^0.4.27"
prometheus-client = "^0.19.0"
prometheus-fastapi-instrumentator = "^6.1.0"
python-consul = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
faker = "^20.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"
isort = "^5.13.0"
psycopg2-binary = "^2.9.9"  # For database setup script

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=html --cov-report=term-missing"
